requirements.txt
aiogram>=3.0
aiohttp
asyncpg
beautifulsoup4
lxml
python-dotenv
